+++ CHANGELOG.md
@@ -6,7 +6,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v0.2.0
+## Unreleased
+
+### Added
+
+- `OrderedLots::remove_by_index()` removes a value by its index.
+- `OrderedLots::index_of_id()` returns the index of a contained LotId.
+- `OrderedLots::id()` returns the LotId for a given index.
+
+## v0.2.0 (2023-06-21)
 
 ### Breaking Changes
 
@@ -23,6 +31,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   the final item from the collection.
 - `LotId` now implements `Ord`.
 
-## v0.1.0
+## v0.1.0 (2023-05-02)
 
 Initial release.